Dump Lsass process with powershell
This mod will allow you to dump the lsass process using powershell.  It will write the dump to a file in c:\windows\temp\(random_name).  Once that is finished it will upload the file to a listening server, and the server can use SSL if selected.
However using SSL is very slow when uploading big files.

Cleaned up the server class a little bit, added new upload with base64.