Style and Functionality Fixes
Swapped y and n in gets
Used vprint instead of print
When a client connects to the server it now prints the clients IP instead of just stating that some one has connected.
Added time out of 90 seconds, gives enought time for the file to be uploaded to the server.