Added Database support
I wrote this to have a centralized location where all the passwords would be stored.

This pull request gives the ability to save passwords taken from the dc and workstations/servers to a database.
smbexec.yml has a new entry for the database name to use
-D was added to smbexec.rb to set the database name during startup
Added a way to interact with the database in the options menu