Multiple DB support
Fully tested with SQLite and Postgres.  The user can choose what db they
would like to use in the yml file.  If the driver is sqlite, the rest of
the information can stay blank.  If the driver is postgres, the user,
pass, host, port and db need to be filled out.  No other code needs to
be changed, importing into the db works, and pulling information out of
the db also works without trouble.